Add a very rough first draft of a PHP 5.6 migration guide just in time for ɑ1.
I've intentionally limited the scope of the changes to the new migration guide:
we shouldn't start documenting new features and changes in the main text of the
manual until after feature freeze, IMO. This is here mostly so that we have
something useful to link to in the news post.

Some of the examples are very, very ropey right now, particularly for new
features. I expect we'll improve them as we go.


git-svn-id: https://svn.php.net/repository/phpdoc/doc-base@332637 c90b9560-bf6c-de11-be94-00142212c4b1